+++ src/component-wrapper/ComponentWrapper.tsx
@@ -2,11 +2,11 @@ import PropTypes from "prop-types";
 import React from "react";
 
 import { Database } from "../database/Database";
-import { NamedSchema, Schema, StoreNames, StoreValue } from "../database/types";
+import { NamedSchema, Schema, StoreNames } from "../database/types";
 
 import { WrappedComponent } from "./internal/WrappedComponent";
 
-import { ComponentDatabaseMap } from "./ComponentDatabaseMap";
+import { ComponentDatabaseMap, ComponentValue } from "./ComponentDatabaseMap";
 import { DynamicComponentType, DynamicComponent } from "./DynamicComponent";
 import { StaticComponent } from "./StaticComponent";
 
@@ -18,7 +18,7 @@ export interface ComponentWrapperRenderProps<
   StoreName extends StoreNames<DBSchema["schema"]>
 > {
   database?: Database<DBSchema>;
-  onChange(value: StoreValue<DBSchema["schema"], StoreName>): void;
+  onChange(value: "" | ComponentValue<DBSchema, StoreName>): void;
 }
 
 /**
@@ -41,15 +41,16 @@ export class ComponentWrapper<
    * The proptype validator for a {@link ComponentWrapper}.
    */
   static readonly propType: PropTypes.Requireable<
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    ComponentWrapper<NamedSchema<string, number, any>, string>
+    ComponentWrapper<NamedSchema<string, number, Schema>, StoreNames<Schema>>
   > = PropTypes.exact({
     key: PropTypes.string.isRequired,
     render: PropTypes.func.isRequired,
     renderWhen: PropTypes.func.isRequired,
     databaseMap: PropTypes.instanceOf<
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      ComponentDatabaseMap<NamedSchema<string, number, any>, string>
+      ComponentDatabaseMap<
+        NamedSchema<string, number, Schema>,
+        StoreNames<Schema>
+      >
     >(ComponentDatabaseMap),
     defaultValue: PropTypes.any,
     emptyValue: PropTypes.any.isRequired
@@ -67,29 +68,13 @@ export class ComponentWrapper<
   static wrapStatic<DBSchema extends NamedSchema<string, number, Schema>>(
     component: StaticComponent<React.ElementType, DBSchema>
   ): ComponentWrapper<DBSchema, StoreNames<DBSchema["schema"]>> {
-    const { key, Component, renderWhen } = component;
+    const { key, Component, props, renderWhen } = component;
 
-    let render: () => JSX.Element;
-
-    if (typeof Component === "string") {
-      const {
-        Component: IntrinsicElement,
-        props
-      } = component as StaticComponent<keyof JSX.IntrinsicElements, DBSchema>;
-
-      // eslint-disable-next-line react/display-name
-      render = (): JSX.Element => <IntrinsicElement key={key} {...props} />;
-    } else {
-      const { Component: ReactComponent, props } = component as StaticComponent<
-        React.ComponentType,
-        DBSchema
-      >;
-
-      // eslint-disable-next-line react/display-name
-      render = (): JSX.Element => <ReactComponent key={key} {...props} />;
-    }
-
-    return new ComponentWrapper(key, render, renderWhen);
+    return new ComponentWrapper(
+      key,
+      (): JSX.Element => <Component key={key} {...props} />,
+      renderWhen
+    );
   }
 
   /**
@@ -102,13 +87,15 @@ export class ComponentWrapper<
   static wrapDynamic<
     Props,
     DBSchema extends NamedSchema<string, number, Schema>,
-    StoreName extends StoreNames<DBSchema["schema"]>
+    StoreName extends StoreNames<DBSchema["schema"]>,
+    Value extends ComponentValue<DBSchema, StoreName>
   >(
     component: DynamicComponent<
-      DynamicComponentType<Props, StoreValue<DBSchema["schema"], StoreName>>,
+      DynamicComponentType<Props, Value>,
       Props,
       DBSchema,
-      StoreName
+      StoreName,
+      Value
     >
   ): ComponentWrapper<DBSchema, StoreName> {
     const {
@@ -152,7 +139,7 @@ export class ComponentWrapper<
    */
   readonly renderWhen: (stepValues: {
     [key: string]:
-      | StoreValue<DBSchema["schema"], StoreNames<DBSchema["schema"]>>
+      | ComponentValue<DBSchema, StoreNames<DBSchema["schema"]>>
       | undefined;
   }) => boolean;
 
@@ -166,12 +153,12 @@ export class ComponentWrapper<
    * The optional default value to store in the {@link Database} if the
    * component hasn't been changed by the user.
    */
-  readonly defaultValue?: StoreValue<DBSchema["schema"], StoreName> | null;
+  readonly defaultValue?: ComponentValue<DBSchema, StoreName> | null;
 
   /**
    * The value to consider as an empty input when updating the {@link Database}.
    */
-  readonly emptyValue: "" | StoreValue<DBSchema["schema"], StoreName>;
+  readonly emptyValue: "" | ComponentValue<DBSchema, StoreName>;
 
   /**
    * Do not use this directly. Use {@link ComponentWrapper.wrapStatic} or
@@ -187,12 +174,12 @@ export class ComponentWrapper<
     ) => JSX.Element,
     renderWhen: (stepValues: {
       [key: string]:
-        | StoreValue<DBSchema["schema"], StoreNames<DBSchema["schema"]>>
+        | ComponentValue<DBSchema, StoreNames<DBSchema["schema"]>>
         | undefined;
     }) => boolean,
     databaseMap?: ComponentDatabaseMap<DBSchema, StoreName>,
-    defaultValue?: StoreValue<DBSchema["schema"], StoreName> | null,
-    emptyValue: "" | StoreValue<DBSchema["schema"], StoreName> = ""
+    defaultValue?: ComponentValue<DBSchema, StoreName> | null,
+    emptyValue: "" | ComponentValue<DBSchema, StoreName> = ""
   ) {
     this.key = key;
     this.render = render;