+++ src/component-wrapper/__mocks__/ComponentWrapper.tsx
@@ -1,14 +1,9 @@
 import React from "react";
 
 import { Database } from "../../database/Database";
-import {
-  NamedSchema,
-  Schema,
-  StoreNames,
-  StoreValue
-} from "../../database/types";
+import { NamedSchema, Schema, StoreNames } from "../../database/types";
 
-import { ComponentDatabaseMap } from "../ComponentDatabaseMap";
+import { ComponentDatabaseMap, ComponentValue } from "../ComponentDatabaseMap";
 import { DynamicComponentType, DynamicComponent } from "../DynamicComponent";
 import { StaticComponent } from "../StaticComponent";
 
@@ -23,7 +18,7 @@ export interface ComponentWrapperRenderProps<
   StoreName extends StoreNames<DBSchema["schema"]>
 > {
   database?: Database<DBSchema>;
-  onChange(value: StoreValue<DBSchema["schema"], StoreName>): void;
+  onChange(value: "" | ComponentValue<DBSchema, StoreName>): void;
 }
 
 export class ComponentWrapper<
@@ -32,8 +27,7 @@ export class ComponentWrapper<
 > extends ActualComponentWrapper<DBSchema, StoreName> {
   static wrapStatic<DBSchema extends NamedSchema<string, number, Schema>>(
     component: StaticComponent<React.ElementType, DBSchema>
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  ): ComponentWrapper<NamedSchema<string, number, any>, string> {
+  ): ComponentWrapper<DBSchema, StoreNames<DBSchema["schema"]>> {
     const { key, Component, renderWhen } = component;
 
     return new ComponentWrapper(
@@ -53,13 +47,15 @@ export class ComponentWrapper<
   static wrapDynamic<
     Props,
     DBSchema extends NamedSchema<string, number, Schema>,
-    StoreName extends StoreNames<DBSchema["schema"]>
+    StoreName extends StoreNames<DBSchema["schema"]>,
+    Value extends ComponentValue<DBSchema, StoreName>
   >(
     component: DynamicComponent<
-      DynamicComponentType<Props, StoreValue<DBSchema["schema"], StoreName>>,
+      DynamicComponentType<Props, Value>,
       Props,
       DBSchema,
-      StoreName
+      StoreName,
+      Value
     >
   ): ComponentWrapper<DBSchema, StoreName> {
     const {
@@ -92,12 +88,12 @@ export class ComponentWrapper<
     ) => JSX.Element,
     renderWhen: (stepValues: {
       [key: string]:
-        | StoreValue<DBSchema["schema"], StoreNames<DBSchema["schema"]>>
+        | ComponentValue<DBSchema, StoreNames<DBSchema["schema"]>>
         | undefined;
     }) => boolean,
     databaseMap?: ComponentDatabaseMap<DBSchema, StoreName>,
-    defaultValue?: StoreValue<DBSchema["schema"], StoreName> | null,
-    emptyValue?: "" | StoreValue<DBSchema["schema"], StoreName>
+    defaultValue?: ComponentValue<DBSchema, StoreName> | null,
+    emptyValue?: "" | ComponentValue<DBSchema, StoreName>
   ) {
     super(key, render, renderWhen, databaseMap, defaultValue, emptyValue);
   }