+++ src/component-wrapper/internal/WrappedComponent.test.tsx
@@ -1,3 +1,4 @@
+import { nullValuesAsUndefined } from "null-as-undefined";
 import React from "react";
 import { ReactTestRenderer, act, create } from "react-test-renderer";
 
@@ -11,7 +12,10 @@ import { DatabaseContext } from "../../database-context/DatabaseContext";
 import { DatabaseProvider } from "../../database-context/DatabaseProvider";
 
 import { ComponentDatabaseMap } from "../ComponentDatabaseMap";
-import { DynamicComponent } from "../DynamicComponent";
+import {
+  DynamicComponent,
+  DynamicComponentControlledProps
+} from "../DynamicComponent";
 
 import { WrappedComponent } from "./WrappedComponent";
 
@@ -24,6 +28,13 @@ type TestSchema = NamedSchema<
     testStore: {
       key: number;
       value: string;
+    };
+
+    anotherTestStore: {
+      key: number;
+      value: {
+        a?: { value?: string };
+      };
     };
   }
 >;
@@ -166,6 +177,136 @@ it("fetches the stored value specified by the `databaseMap` when a database prov
         disabled={false}
         onChange={[Function]}
         value="testStore/0"
+      />
+    </div>
+  `);
+});
+
+it("fetches the stored value specified by the `databaseMap` and uses the specified child property", async () => {
+  const get = spyOnDatabaseGet(true, { a: { value: "new value" } });
+
+  const Component = (
+    props: DynamicComponentControlledProps<
+      TestSchema["schema"]["anotherTestStore"]["value"]["a"]
+    >
+  ): JSX.Element => {
+    const { value, onValueChange, disabled } = nullValuesAsUndefined(props);
+
+    return (
+      <input
+        value={JSON.stringify(value)}
+        onChange={(event: React.ChangeEvent<HTMLInputElement>): void => {
+          // This is bad practice in real situations. Don't allow a user to
+          // modify the object themselves.
+          onValueChange(JSON.parse(event.target.value));
+        }}
+        disabled={disabled}
+      />
+    );
+  };
+
+  const databaseMap = new ComponentDatabaseMap<TestSchema, "anotherTestStore">({
+    storeName: "anotherTestStore",
+    key: 0,
+    property: "a"
+  });
+
+  const dynamicComponent = new DynamicComponent({
+    key: "test-component",
+    Component,
+    props: {},
+    defaultValue: {},
+    databaseMap
+  });
+
+  const openPromise = Database.open<TestSchema>("testDBName", 1);
+  const DBContext = new DatabaseContext(openPromise);
+
+  let component: ReactTestRenderer | undefined = undefined;
+
+  await act(async () => {
+    component = create(
+      <DatabaseProvider context={DBContext}>
+        <DBContext.Consumer>
+          {(database): JSX.Element => (
+            <WrappedComponent
+              database={database}
+              component={dynamicComponent}
+            />
+          )}
+        </DBContext.Consumer>
+      </DatabaseProvider>
+    );
+
+    await openPromise;
+    await get.settle;
+  });
+
+  expect(get.spy).toHaveBeenCalledTimes(1);
+
+  expect(component).toMatchInlineSnapshot(`
+    <input
+      disabled={false}
+      onChange={[Function]}
+      value="{\\"value\\":\\"new value\\"}"
+    />
+  `);
+});
+
+it("fetches the stored value specified by the `databaseMap` and uses the specified child property path", async () => {
+  const get = spyOnDatabaseGet(true, { a: { value: "new value" } });
+
+  const databaseMap = new ComponentDatabaseMap<TestSchema, "anotherTestStore">({
+    storeName: "anotherTestStore",
+    key: 0,
+    property: ["a", "value"]
+  });
+
+  const dynamicComponent = new DynamicComponent({
+    key: "test-component",
+    Component: TestDynamicComponent,
+    props: {
+      content: "test content"
+    },
+    defaultValue: "test default value",
+    databaseMap
+  });
+
+  const openPromise = Database.open<TestSchema>("testDBName", 1);
+  const DBContext = new DatabaseContext(openPromise);
+
+  let component: ReactTestRenderer | undefined = undefined;
+
+  await act(async () => {
+    component = create(
+      <DatabaseProvider context={DBContext}>
+        <DBContext.Consumer>
+          {(database): JSX.Element => (
+            <WrappedComponent
+              database={database}
+              component={dynamicComponent}
+            />
+          )}
+        </DBContext.Consumer>
+      </DatabaseProvider>
+    );
+
+    await openPromise;
+    await get.settle;
+  });
+
+  expect(get.spy).toHaveBeenCalledTimes(1);
+
+  expect(component).toMatchInlineSnapshot(`
+    <div>
+      <div>
+        test content
+      </div>
+      <input
+        data-testid="input"
+        disabled={false}
+        onChange={[Function]}
+        value="new value"
       />
     </div>
   `);