+++ src/database/types.ts
@@ -1,17 +1,55 @@
+import { IDBPTransaction, StoreNames as DBStoreNames } from "idb";
+
 import {
-  IDBPObjectStore,
-  IDBPTransaction,
-  StoreKey as DBStoreKey,
-  StoreNames as DBStoreNames,
-  StoreValue as DBStoreValue
-} from "idb";
+  BooleanProperties,
+  NumberProperties,
+  StringProperties
+} from "../helpers/internal/primitives";
+
+type StorePropertyType = string | number | boolean | Date;
+
+type StorePropertyTypeProperties =
+  | StringProperties
+  | NumberProperties
+  | BooleanProperties
+  | keyof Date
+  | keyof StorePropertyType[];
 
 /**
- * The schema for an individual store.
- *
- * Stores are semantically similar to tables in other databases.
+ * The type `T` without its inherent primitive prototype properties.
  */
-export interface StoreSchema<Value> {
+export type StoreValueProperties<T> = Omit<T, StorePropertyTypeProperties>;
+
+/**
+ * The type of a possible value in a {@link StoreSchema}.
+ */
+export type StoreSchemaValue =
+  | StorePropertyType
+  | StorePropertyType[]
+  | {
+      [s: string]:
+        | StorePropertyType
+        | { [s: string]: StorePropertyType | undefined }
+        | { [n: number]: StorePropertyType | undefined }
+        | undefined;
+    }
+  | {
+      [n: number]:
+        | StorePropertyType
+        | { [s: string]: StorePropertyType | undefined }
+        | { [n: number]: StorePropertyType | undefined }
+        | undefined;
+    };
+
+/**
+ * The schema for an individual {@link Store}.
+ *
+ * {@link Store|Stores} are semantically similar to tables in other databases.
+ *
+ * You can store objects in {@link Store|Stores}, but only one with a
+ * maximum of two levels of depth.
+ */
+export interface StoreSchema<Value extends StoreSchemaValue> {
   /**
    * The type of keys used by the store.
    */
@@ -34,8 +72,7 @@ export interface StoreSchema<Value> {
  * IndexedDB databases are made up of stores, mapping store name to
  * {@link StoreSchema}.
  */
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export interface Schema<Value = any> {
+export interface Schema<Value extends StoreSchemaValue = StoreSchemaValue> {
   [storeName: string]: StoreSchema<Value>;
 }
 
@@ -102,7 +139,7 @@ export type StoreNames<DBSchema extends Schema> = DBStoreNames<DBSchema> &
 export type StoreKey<
   DBSchema extends Schema,
   StoreName extends StoreNames<DBSchema>
-> = DBStoreKey<DBSchema, StoreName>;
+> = DBSchema[StoreName]["key"];
 
 /**
  * The type of values in the given store.
@@ -110,7 +147,59 @@ export type StoreKey<
 export type StoreValue<
   DBSchema extends Schema,
   StoreName extends StoreNames<DBSchema>
-> = DBStoreValue<DBSchema, StoreName>;
+> = DBSchema[StoreName]["value"];
+
+// ["a"] | ["z"]
+// This follows the same pattern as `ComponentValueLevelOne`.
+type StoreValuePropertyPathLevelOne<Value extends {}> = {
+  [K in keyof StoreValueProperties<Value>]: [K];
+}[keyof StoreValueProperties<Value>];
+
+// ["a", "b"] | ["z", "y"]
+// This follows the same pattern as `ComponentValueLevelTwo`.
+type StoreValuePropertyPathLevelTwo<Value extends {}> = {
+  [K in keyof StoreValueProperties<Value>]: keyof StoreValueProperties<
+    NonNullable<StoreValueProperties<Value>[K]>
+  > extends never // If `Value[K]` is a primitive...
+    ? never // ...ignore it...
+    : [
+        K,
+        keyof StoreValueProperties<NonNullable<StoreValueProperties<Value>[K]>>
+      ]; // ...otherwise, allow referencing its children.
+}[keyof StoreValueProperties<Value>];
+
+/**
+ * The type describing the "path" to a deep property of a {@link StoreValue}.
+ *
+ * The resulting type is a union of arrays describing all valid "paths" to deep
+ * properties.
+ *
+ * This only goes two levels deep, at the most, including the base level. If
+ * your {@link Schema} has deeper nesting than that and it needs to connect a
+ * {@link DynamicComponent} to a deep property, consider refactoring. More than
+ * two levels implies a schema of some complexity that will be difficult to
+ * maintain.
+ *
+ * For example
+ *
+ * ```ts
+ * type PropertyPath = StoreValuePropertyPath<{
+ *   a: { b: boolean };
+ *   z: number;
+ * }>;
+ * ```
+ *
+ * is the same as
+ *
+ * ```ts
+ * type PropertyPath =
+ *   | ["a"] | ["z"]
+ *   | ["a", "b"];
+ * ```
+ */
+export type StoreValuePropertyPath<Value extends {}> = NonNullable<
+  StoreValuePropertyPathLevelOne<Value> | StoreValuePropertyPathLevelTwo<Value>
+>;
 
 /**
  * A transaction on one or more stores.
@@ -119,31 +208,3 @@ export type Transaction<
   DBSchema extends Schema,
   TxStoreNames extends StoreNames<DBSchema>[] = StoreNames<DBSchema>[]
 > = IDBPTransaction<DBSchema, TxStoreNames>;
-
-/**
- * A store accessor.
- *
- * Stores are semantically similar to tables in other databases.
- */
-export type Store<
-  DBSchema extends Schema,
-  TxStoreNames extends StoreNames<DBSchema>[] = StoreNames<DBSchema>[],
-  StoreName extends StoreNames<DBSchema> = StoreNames<DBSchema>
-> = IDBPObjectStore<DBSchema, TxStoreNames, StoreName>;
-
-/**
- * A map of the name of a {@link Store} to its accessor.
- *
- * @typeparam TxStoreNames - The array of the names of stores possibly in this
- * map.
- */
-export type StoreMap<
-  DBSchema extends Schema,
-  TxStoreNames extends StoreNames<DBSchema>[] = StoreNames<DBSchema>[]
-> = {
-  [Name in TxStoreNames extends (infer N)[] ? N : never]: Store<
-    DBSchema,
-    TxStoreNames,
-    Name
-  >;
-};