+++ src/helpers/internal/object.ts
@@ -1,0 +1,5 @@
+export type ChildOf<T> = T extends {}
+  ? {
+      [K in keyof NonNullable<T>]: NonNullable<T>[K];
+    }[keyof NonNullable<T>]
+  : never;