+++ src/step/internal/Step.test.tsx
@@ -1,4 +1,4 @@
-/* eslint-disable @typescript-eslint/no-empty-function, @typescript-eslint/unbound-method */
+/* eslint-disable @typescript-eslint/no-empty-function */
 import { render, fireEvent } from "@testing-library/react";
 import React from "react";
 import { create } from "react-test-renderer";
@@ -15,6 +15,7 @@ import {
   spyOnDatabaseGet,
   spyOnDatabaseTransaction
 } from "../../__tests__/helpers/spies";
+import { promiseToWaitForNextTick } from "../../__tests__/helpers/promise";
 
 import { ComponentDatabaseMap } from "../../component-wrapper/ComponentDatabaseMap";
 import { ComponentWrapper } from "../../component-wrapper/ComponentWrapper";
@@ -26,6 +27,7 @@ import { DatabaseContext } from "../../database-context/DatabaseContext";
 import { DatabaseProvider } from "../../database-context/DatabaseProvider";
 
 import { Step } from "./Step";
+import { Store } from "../../database/Store";
 
 jest.mock("../../component-wrapper/ComponentWrapper");
 jest.mock("../../database/Database");
@@ -221,6 +223,7 @@ it("persists data to the database when submitted", async () => {
 
   const get = spyOnDatabaseGet();
   const transaction = spyOnDatabaseTransaction();
+  const storePut = jest.spyOn(Store.prototype, "put");
 
   type TestSchema = NamedSchema<
     string,
@@ -267,6 +270,7 @@ it("persists data to the database when submitted", async () => {
   );
 
   await get.settle;
+  await promiseToWaitForNextTick();
 
   const newValue = "new value";
 
@@ -282,12 +286,8 @@ it("persists data to the database when submitted", async () => {
 
   await transaction.settle;
 
-  const store = transaction.stores[databaseMap.storeName];
-
-  expect(store).toBeDefined();
-
-  expect(store.put).toHaveBeenCalledTimes(1);
-  expect(store.put).toHaveBeenCalledWith(newValue, databaseMap.key);
+  expect(storePut).toHaveBeenCalledTimes(1);
+  expect(storePut).toHaveBeenCalledWith(databaseMap.key, newValue);
 });
 
 it("deletes the corresponding data from the database when submitted with the empty value", async () => {
@@ -297,6 +297,7 @@ it("deletes the corresponding data from the database when submitted with the emp
 
   const get = spyOnDatabaseGet();
   const transaction = spyOnDatabaseTransaction();
+  const storeDelete = jest.spyOn(Store.prototype, "delete");
 
   type TestSchema = NamedSchema<
     string,
@@ -346,6 +347,7 @@ it("deletes the corresponding data from the database when submitted with the emp
   );
 
   await get.settle;
+  await promiseToWaitForNextTick();
 
   fireEvent.change(getByTestId("input"), { target: { value: emptyValue } });
   fireEvent.click(getByTestId("submit"));
@@ -359,10 +361,158 @@ it("deletes the corresponding data from the database when submitted with the emp
 
   await transaction.settle;
 
-  const store = transaction.stores[databaseMap.storeName];
+  expect(storeDelete).toHaveBeenCalledTimes(1);
+  expect(storeDelete).toHaveBeenCalledWith(databaseMap.key);
+});
 
-  expect(store).toBeDefined();
+it("persists child property data to the database when submitted", async () => {
+  const { ComponentWrapper } = jest.requireActual(
+    "../../component-wrapper/ComponentWrapper"
+  );
 
-  expect(store.delete).toHaveBeenCalledTimes(1);
-  expect(store.delete).toHaveBeenCalledWith(databaseMap.key);
+  const get = spyOnDatabaseGet(true, { a: { value: "old value" } });
+  const transaction = spyOnDatabaseTransaction();
+  const storePut = jest.spyOn(Store.prototype, "put");
+
+  type TestSchema = NamedSchema<
+    string,
+    number,
+    {
+      testStore: {
+        key: number;
+        value: {
+          a: { value: string };
+        };
+      };
+    }
+  >;
+
+  const databaseMap = new ComponentDatabaseMap<TestSchema, "testStore">({
+    storeName: "testStore",
+    key: 0,
+    property: ["a", "value"]
+  });
+
+  const wrappers = [
+    ComponentWrapper.wrapDynamic(
+      new DynamicComponent({
+        key: "test-component",
+        Component: TestDynamicComponent,
+        props: {
+          key: "test-key",
+          content: "test content"
+        },
+        defaultValue: "test default value",
+        databaseMap
+      })
+    )
+  ];
+
+  const database = await Database.open("testDBName", 1);
+  const DBContext = new DatabaseContext(database);
+
+  const { getByTestId } = render(
+    <DatabaseProvider context={DBContext} allowUnmounting>
+      <Step
+        context={DBContext}
+        componentWrappers={wrappers}
+        Submit={SimpleSubmit}
+      />
+    </DatabaseProvider>
+  );
+
+  await get.settle;
+  await promiseToWaitForNextTick();
+
+  const newValue = "new value";
+
+  fireEvent.change(getByTestId("input"), { target: { value: newValue } });
+  fireEvent.click(getByTestId("submit"));
+
+  expect(transaction.spy).toHaveBeenCalledTimes(1);
+
+  await transaction.settle;
+
+  expect(storePut).toHaveBeenCalledTimes(1);
+  expect(storePut).toHaveBeenCalledWith(databaseMap.key, {
+    a: { value: newValue }
+  });
+});
+
+it("removes the corresponding child property data from the database when submitted with the empty value", async () => {
+  const { ComponentWrapper } = jest.requireActual(
+    "../../component-wrapper/ComponentWrapper"
+  );
+
+  const oldValue = { a: { value: "old value" } };
+
+  const get = spyOnDatabaseGet(true, oldValue);
+  const transaction = spyOnDatabaseTransaction();
+  const storeGet = jest.spyOn(Store.prototype, "get");
+  const storePut = jest.spyOn(Store.prototype, "put");
+
+  storeGet.mockResolvedValue(oldValue);
+
+  type TestSchema = NamedSchema<
+    string,
+    number,
+    {
+      testStore: {
+        key: number;
+        value: {
+          a: { value: string };
+        };
+      };
+    }
+  >;
+
+  const databaseMap = new ComponentDatabaseMap<TestSchema, "testStore">({
+    storeName: "testStore",
+    key: 0,
+    property: ["a", "value"]
+  });
+
+  const emptyValue = "test empty value";
+
+  const wrappers = [
+    ComponentWrapper.wrapDynamic(
+      new DynamicComponent({
+        key: "test-component",
+        Component: TestDynamicComponent,
+        props: {
+          key: "test-key",
+          content: "test content"
+        },
+        defaultValue: "test default value",
+        emptyValue,
+        databaseMap
+      })
+    )
+  ];
+
+  const database = await Database.open("testDBName", 1);
+  const DBContext = new DatabaseContext(database);
+
+  const { getByTestId } = render(
+    <DatabaseProvider context={DBContext} allowUnmounting>
+      <Step
+        context={DBContext}
+        componentWrappers={wrappers}
+        Submit={SimpleSubmit}
+      />
+    </DatabaseProvider>
+  );
+
+  await get.settle;
+  await promiseToWaitForNextTick();
+
+  fireEvent.change(getByTestId("input"), { target: { value: emptyValue } });
+  fireEvent.click(getByTestId("submit"));
+
+  expect(transaction.spy).toHaveBeenCalledTimes(1);
+
+  await transaction.settle;
+
+  expect(storePut).toHaveBeenCalledTimes(1);
+  expect(storePut).toHaveBeenCalledWith(databaseMap.key, { a: {} });
 });